Flask/backend/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-cors==3.0.10
pyjwt==2.1.0
python-dotenv==0.19.0
scikit-learn==0.24.2
joblib==1.0.1
flask-socketio==5.1.1
eventlet==0.33.0
pandas==1.3.5
numpy==1.21.6
passlib==1.7.4